requirements.txt
Django==3.2.8
django-allauth==0.45.0
Pillow==8.3.2
